readd file removed by mistake (and the remove the that that should)
