--- Polyhedron/demo/Polyhedron/Plugins/Mesh_3/Volume_planes_plugin.cpp
@@ -1,663 +1,0 @@
-
-#ifdef _MSC_VER
-#  pragma warning(disable:4244) // conversion with loss of data
-#endif
-
-#include <CGAL/config.h>
-#include "config.h"
-
-#include "Volume_plane.h"
-#include <CGAL/Image_3.h>
-
-#include "Volume_plane_thread.h"
-#include "Volume_plane_intersection.h"
-#include "Scene_segmented_image_item.h"
-#include <CGAL/Three/Polyhedron_demo_plugin_interface.h>
-#include "Messages_interface.h"
-#include <CGAL/Three/Scene_interface.h>
-#include <CGAL/Three/Scene_item.h>
-#include <CGAL/Three/Viewer_interface.h>
-#include <CGAL/Three/Scene_group_item.h>
-
-#include <QAction>
-#include <QMenu>
-#include <QList>
-#include <QInputDialog>
-#include <QSlider>
-#include <QLabel>
-#include <QHBoxLayout>
-#include <QVBoxLayout>
-#include <QDockWidget>
-#include <QMainWindow>
-#include <QMessageBox>
-#include <QString>
-#include <QFontMetrics>
-#include <QMouseEvent>
-
-#include <cassert>
-#include <iostream>
-
-#include <boost/type_traits.hpp>
-#include <boost/optional.hpp>
-#include "Scene.h"
-// Covariant return types don't work for scalar types and we cannot
-// have templates here, hence this unfortunate hack.
-
-// The input float value we are reading is always in
-// 0..1 and min_max is the range it came from.
-struct IntConverter {
-  std::pair<int, int> min_max;
-  
-  int operator()(float f) {
-    float s = f * (min_max.second - min_max.first);
-    return s + min_max.first;
-  }
-};
-
-struct DoubleConverter {
-  std::pair<float, float> min_max;
-
-  float operator()(float f) {
-    float s = f * (min_max.second - min_max.first);
-    return s + min_max.first;
-  }
-};
-
-class PixelReader : public QObject
-{
-Q_OBJECT
-public Q_SLOTS:
-  void update(const QMouseEvent *e) {
-    getPixel(e->pos());
-  }
-Q_SIGNALS:
-  void x(int);
-
-public:
-  void setIC(const IntConverter& x) { ic = x; fc = boost::optional<DoubleConverter>(); }
-  void setFC(const DoubleConverter& x) { fc = x; ic = boost::optional<IntConverter>(); }
-  void setViewer(Viewer_interface* viewer) { this->viewer = viewer; }
-
-private:
-  boost::optional<IntConverter> ic;
-  boost::optional<DoubleConverter> fc;
-  Viewer_interface* viewer;
-
-  void getPixel(const QPoint& e) {
-    float data[3];
-    int vp[4];
-    viewer->glGetIntegerv(GL_VIEWPORT, vp);
-    viewer->glReadPixels(e.x(), vp[3] - e.y(), 1, 1, GL_RGB, GL_FLOAT, data);
-
-    if(fc) {
-      Q_EMIT x( (*fc)(data[0]) );
-    } else if(ic) {
-      Q_EMIT x( (*ic)(data[0]) );
-    }
-  }
-};
-
-
-class Plane_slider : public QSlider
-{
-  Q_OBJECT
-public:
-  Plane_slider(const qglviewer::Vec& v, int id, Scene_interface* scene,
-               qglviewer::ManipulatedFrame* frame, Qt::Orientation ori, QWidget* widget) 
-    : QSlider(ori, widget), v(v), id(id), scene(scene), frame(frame) { 
-    this->setTracking(true);
-    connect(frame,  SIGNAL(manipulated()), this, SLOT(updateValue()));
-  }
-
-public:
-  void sliderChange(SliderChange c) {
-    QSlider::sliderChange(c);
-    if(c == SliderValueChange) {
-      qglviewer::Vec v2 = v * (this->value() / scale);
-      frame->setTranslationWithConstraint(v2);
-      scene->itemChanged(id);
-    }
-
-    Q_EMIT realChange(this->value() / scale);
-  }
-
-public Q_SLOTS:
-  void updateValue() {
-#if QGLVIEWER_VERSION >= 0x020600
-    typedef qreal qglviewer_real;
-#else // QGLViewer < 2.6.0
-    typedef float qglviewer_real;
-#endif // QGLViewer < 2.6.0
-    qglviewer_real a, b, c;
-    frame->getPosition(a, b, c);
-    float sum1 = float(a + b + c);
-    float sum2 = float(v.x + v.y + v.z);
-    sum1 /= sum2;
-    setValue(sum1 * scale);
-  }
-
-Q_SIGNALS:
-  void realChange(int);
-
-private:
-  static const unsigned int scale;
-
-  qglviewer::Vec v;
-  int id;
-  Scene_interface* scene;
-  qglviewer::ManipulatedFrame* frame;
-};
-
-const unsigned int Plane_slider::scale = 100;
-
-class Volume_plane_plugin :
-  public QObject,
-  public CGAL::Three::Polyhedron_demo_plugin_interface
-{
-  Q_OBJECT
-  Q_INTERFACES(CGAL::Three::Polyhedron_demo_plugin_interface)
-  Q_PLUGIN_METADATA(IID "com.geometryfactory.PolyhedronDemo.PluginInterface/1.0")
-
-public:
-  Volume_plane_plugin() : planeSwitch(NULL)
-    {
-    }
-
-  bool applicable(QAction*) const {
-    return qobject_cast<Scene_segmented_image_item*>(scene->item(scene->mainSelectionIndex()));
-  }
-
-  void init(QMainWindow* mainWindow, CGAL::Three::Scene_interface* scene_interface, Messages_interface*) {
-    this->scene = scene_interface;
-    this->mw = mainWindow;
-    x_control = NULL;
-    y_control = NULL;
-    z_control = NULL;
-    current_control = NULL;
-    planeSwitch = new QAction("Add Volume Planes", mw);
-    if(planeSwitch) {
-      planeSwitch->setProperty("subMenuName", "3D Mesh Generation");
-      connect(planeSwitch, SIGNAL(triggered()),
-              this, SLOT(selectPlanes()));
-      Scene* true_scene = dynamic_cast<Scene*>(scene);
-      connect(true_scene,SIGNAL(selectionChanged(int)),
-              this, SLOT(connect_controls(int)));
-    }
-    Viewer_interface* v = mw->findChild<Viewer_interface*>("viewer");
-    CGAL_assertion(v != 0);
-    pxr_.setViewer(v);
-    connect(v, SIGNAL(pointSelected(const QMouseEvent *)), &pxr_, SLOT(update(const QMouseEvent *)));
-    createOrGetDockLayout();
-
-  }
-  QList<QAction*> actions() const {
-    return QList<QAction*>() << planeSwitch;
-  }
-  virtual void closure()
-  {
-      QDockWidget* controlDockWidget = mw->findChild<QDockWidget*>("volumePlanesControl");
-      if(controlDockWidget)
-          controlDockWidget->hide();
-  }
-public Q_SLOTS:
-  void selectPlanes() {
-
-    //Control widgets creation
-    QLayout* layout = createOrGetDockLayout();
-    if(x_control == NULL)
-    {
-      x_control = new QWidget;
-      x_box = new QHBoxLayout(x_control);
-      layout->addWidget(x_control);
-
-      QLabel* label = new QLabel(x_control);
-      label->setText("X Slice");
-
-      x_cubeLabel = new QLabel(x_control);
-
-      // Find the right width for the label to accommodate at least 9999
-      QFontMetrics metric = x_cubeLabel->fontMetrics();
-      x_cubeLabel->setFixedWidth(metric.width(QString("9999")));
-      x_cubeLabel->setNum(0);
-      x_slider = new QSlider(mw);
-
-      x_box->addWidget(label);
-      x_box->addWidget(x_slider);
-      x_box->addWidget(x_cubeLabel);
-    }
-
-    if(y_control == NULL)
-    {
-      y_control = new QWidget;
-      y_box = new QHBoxLayout(y_control);
-      layout->addWidget(y_control);
-
-      QLabel* label = new QLabel(y_control);
-      label->setText("Y Slice");
-
-      y_cubeLabel = new QLabel(y_control);
-
-      // Find the right width for the label to accommodate at least 9999
-      QFontMetrics metric = y_cubeLabel->fontMetrics();
-      y_cubeLabel->setFixedWidth(metric.width(QString("9999")));
-      y_cubeLabel->setNum(0);
-
-      y_slider = new QSlider(mw);
-
-      y_box->addWidget(label);
-      y_box->addWidget(y_slider);
-      y_box->addWidget(y_cubeLabel);
-    }
-
-    if(z_control == NULL)
-    {
-      z_control = new QWidget;
-      z_box = new QHBoxLayout(z_control);
-      layout->addWidget(z_control);
-
-      QLabel* label = new QLabel(z_control);
-      label->setText("Z Slice");
-
-      z_cubeLabel = new QLabel(z_control);
-
-      // Find the right width for the label to accommodate at least 9999
-      QFontMetrics metric = z_cubeLabel->fontMetrics();
-      z_cubeLabel->setFixedWidth(metric.width(QString("9999")));
-      z_cubeLabel->setNum(0);
-      z_slider = new QSlider(mw);
-
-      z_box->addWidget(label);
-      z_box->addWidget(z_slider);
-      z_box->addWidget(z_cubeLabel);
-    }
-    std::vector< Scene_segmented_image_item* > seg_items;
-    Scene_segmented_image_item* seg_img = NULL;
-    for(int i = 0; i < scene->numberOfEntries(); ++i) {
-      Scene_segmented_image_item* tmp = qobject_cast<Scene_segmented_image_item*>(scene->item(i));
-      if(tmp != NULL){
-        seg_items.push_back(tmp);
-      }
-    }
-    if(seg_items.empty()) {
-      QMessageBox::warning(mw, tr("No suitable item found"), tr("Load an inrimage or hdr file to enable Volume Planes."));
-      return;
-    } else {
-      QList<QString> items;
-      for(std::vector< Scene_segmented_image_item* >::const_iterator it = seg_items.begin(); 
-          it != seg_items.end(); ++it) { 
-        items << (*it)->name();
-      }
-      bool ok;
-      QString selected = QInputDialog::getItem(mw, tr("Select a dataset:"), tr("Items"), items, 0, false, &ok);
-      if(!ok || selected.isEmpty())
-        return;
-      // Opens a modal Dialog to prevent the user from manipulating things that could mess with the planes creation and cause a segfault.
-      msgBox.setText("Planes created : 0/3");
-      msgBox.setStandardButtons(QMessageBox::NoButton);
-      msgBox.show();
-      for(std::vector< Scene_segmented_image_item*>::const_iterator it = seg_items.begin(); 
-          it != seg_items.end(); ++it) { 
-        if(selected == (*it)->name())
-          seg_img = *it;
-      }
-    }
-    
-    if(!(seg_img == NULL)) {
-      const CGAL::Image_3* img = seg_img->image();
-      CGAL_IMAGE_IO_CASE(img->image(), this->launchAdders<Word>(seg_img, seg_img->name()))
-
-      Volume_plane_intersection* i = new Volume_plane_intersection(img->xdim() * img->vx(), 
-                                                                   img->ydim() * img->vy(), 
-                                                                   img->zdim() * img->vz());
-      this->intersectionId = scene->addItem(i);
-      scene->changeGroup(i, group);
-    } else {
-      QMessageBox::warning(mw, tr("Something went wrong"), tr("Selected a suitable Object but couldn't get an image pointer."));
-      return;
-    }
-
-  }
-  void addVP(Volume_plane_thread* thread) {
-    Volume_plane_interface* plane = thread->getItem();
-    plane->init();
-    // add the interface for this Volume_plane
-    int id = scene->addItem(plane);
-    scene->changeGroup(plane, group);
-    switch(thread->type())
-    {
-    case 'x':
-      delete x_slider;
-      x_slider = new Plane_slider(plane->translationVector(), id, scene, plane->manipulatedFrame(),
-                                         Qt::Horizontal, x_control);
-      x_slider->setRange(0, (plane->cDim() - 1) * 100);
-      connect(x_slider, SIGNAL(realChange(int)), x_cubeLabel, SLOT(setNum(int)));
-      connect(x_slider, SIGNAL(realChange(int)), this, SLOT(set_value()));
-      connect(plane, SIGNAL(manipulated(int)), x_cubeLabel, SLOT(setNum(int)));
-      connect(plane, SIGNAL(aboutToBeDestroyed()), this, SLOT(destroy_x_item()));
-      x_box->addWidget(x_slider);
-      x_box->addWidget(x_cubeLabel);
-      break;
-    case 'y':
-      delete y_slider;
-      y_slider = new Plane_slider(plane->translationVector(), id, scene, plane->manipulatedFrame(),
-                                         Qt::Horizontal, y_control);
-      y_slider->setRange(0, (plane->cDim() - 1) * 100);
-      connect(y_slider, SIGNAL(realChange(int)), y_cubeLabel, SLOT(setNum(int)));
-      connect(y_slider, SIGNAL(realChange(int)), this, SLOT(set_value()));
-      connect(plane, SIGNAL(manipulated(int)), y_cubeLabel, SLOT(setNum(int)));
-      connect(plane, SIGNAL(aboutToBeDestroyed()), this, SLOT(destroy_y_item()));
-      y_box->addWidget(y_slider);
-      y_box->addWidget(y_cubeLabel);
-      break;
-    case 'z':
-      delete z_slider;
-      z_slider = new Plane_slider(plane->translationVector(), id, scene, plane->manipulatedFrame(),
-                                         Qt::Horizontal, z_control);
-      z_slider->setRange(0, (plane->cDim() - 1) * 100);
-      connect(z_slider, SIGNAL(realChange(int)), z_cubeLabel, SLOT(setNum(int)));
-      connect(z_slider, SIGNAL(realChange(int)), this, SLOT(set_value()));
-      connect(plane, SIGNAL(manipulated(int)), z_cubeLabel, SLOT(setNum(int)));
-      connect(plane, SIGNAL(aboutToBeDestroyed()), this, SLOT(destroy_z_item()));
-      z_box->addWidget(z_slider);
-      z_box->addWidget(z_cubeLabel);
-      break;
-    default:
-      break;
-    }
-    std::vector<Volume_plane_thread*>::iterator it = std::find(threads.begin(), threads.end(), thread);
-
-    // this slot has been connected to a thread that hasn't been
-    // registered here.
-    assert(it != threads.end());
-    delete *it;
-    threads.erase(it);
-
-    update_msgBox();
-    Volume_plane_intersection* intersection = dynamic_cast<Volume_plane_intersection*>(scene->item(intersectionId));
-    if(!intersection) {
-      // the intersection is gone before it was initialized
-      return;
-    }
-    // FIXME downcasting mode
-    // FIXME this will bug if two volume planes are generated simultaneously by the plugin
-    if(Volume_plane<x_tag>* p = dynamic_cast< Volume_plane<x_tag>* >(plane)) {
-      intersection->setX(p);
-    } else if(Volume_plane<y_tag>* p = dynamic_cast< Volume_plane<y_tag>* >(plane)) {
-      intersection->setY(p);
-    } else if(Volume_plane<z_tag>* p = dynamic_cast< Volume_plane<z_tag>* >(plane)) {
-      intersection->setZ(p);
-    }    
-    connect(plane, SIGNAL(planeDestructionIncoming(Volume_plane_interface*)), 
-            intersection, SLOT(planeRemoved(Volume_plane_interface*)));
-
-  }
- 
-private:
-  QMessageBox msgBox;
-  QAction* planeSwitch;
-  QWidget *x_control, *y_control, *z_control;
-  QSlider *x_slider, *y_slider, *z_slider;
-  QLabel *x_cubeLabel, *y_cubeLabel, *z_cubeLabel;
-  QHBoxLayout *x_box, *y_box, *z_box;
-  PixelReader pxr_;
-  CGAL::Three::Scene_group_item* group;
-  CGAL::Three::Scene_interface* scene;
-  QMainWindow* mw;
-  std::vector<Volume_plane_thread*> threads;
-  struct controls{
-    CGAL::Three::Scene_item* group;
-    CGAL::Three::Scene_item* xitem;
-    CGAL::Three::Scene_item* yitem;
-    CGAL::Three::Scene_item* zitem;
-    int xvalue;
-    int yvalue;
-    int zvalue;
-  };
-  controls *current_control;
-  QMap<CGAL::Three::Scene_item*, controls> group_map;
-  unsigned int intersectionId;
-
-  QLayout* createOrGetDockLayout() {
-    QLayout* layout = NULL;
-    QDockWidget* controlDockWidget = mw->findChild<QDockWidget*>("volumePlanesControl");;
-
-    if(!controlDockWidget) {
-      controlDockWidget = new QDockWidget(mw);
-      controlDockWidget->setObjectName("volumePlanesControl");
-      QWidget* content = new QWidget(controlDockWidget);
-      layout = new QVBoxLayout(content);
-      layout->setObjectName("vpSliderLayout");
-      controlDockWidget->setWindowTitle("Control Widget");
-      mw->addDockWidget(Qt::LeftDockWidgetArea, controlDockWidget);
-      
-      QWidget* vlabels = new QWidget(content);
-      layout->addWidget(vlabels);
-      QHBoxLayout* vbox = new QHBoxLayout(vlabels);
-      vbox->setAlignment(Qt::AlignJustify);
-      QLabel* help = new QLabel(vlabels);
-      help->setText("Cut planes for the \nselected image:");
-      QLabel* text = new QLabel(vlabels);
-      text->setText("Isovalue at point:");
-      QLabel* x = new QLabel(vlabels);
-
-      connect(&pxr_, SIGNAL(x(int)), x, SLOT(setNum(int)));
-      
-      layout->addWidget(help); vbox->addWidget(text); vbox->addWidget(x);
-      controlDockWidget->setWidget(content);
-      controlDockWidget->hide();
-
-    } else {
-      layout = controlDockWidget->findChild<QLayout*>("vpSliderLayout");
-      controlDockWidget->show();
-    }
-
-    return layout;
-  }
-
-  template<typename Word>
-  void launchAdders(Scene_segmented_image_item* seg_img, const QString& name) {
-    const CGAL::Image_3* img = seg_img->image();
-    const Word* begin = (const Word*)img->data();
-    const Word* end = (const Word*)img->data() + img->size();
-
-    std::pair<const Word, const Word> minmax(*std::min_element(begin, end), *std::max_element(begin, end));
-
-    Clamp_to_one_zero_range clamper = { minmax };
-
-    switchReaderConverter< Word >(minmax);
-
-    Volume_plane<x_tag> *xitem = new Volume_plane<x_tag>();
-    Volume_plane<y_tag> *yitem = new Volume_plane<y_tag>();
-    Volume_plane<z_tag> *zitem = new Volume_plane<z_tag>();
-    scene->setSelectedItem(-1);
-    group = new Scene_group_item(QString("Planes for %1").arg(seg_img->name()));
-    connect(group, SIGNAL(aboutToBeDestroyed()),
-            this, SLOT(erase_group()));
-    scene->addItem(group);
-    controls c;
-    c.group = group;
-    c.xitem = xitem;
-    c.yitem = yitem;
-    c.zitem = zitem;
-    c.xvalue = 0;
-    c.yvalue = 0;
-    c.zvalue = 0;
-    group_map[seg_img] = c;
-    current_control = &group_map[seg_img];
-    connect(seg_img, SIGNAL(aboutToBeDestroyed()),
-            this, SLOT(erase_group()));
-
-    threads.push_back(new X_plane_thread<Word>(xitem, img, clamper, name));
-
-    connect(threads.back(), SIGNAL(finished(Volume_plane_thread*)), this, SLOT(addVP(Volume_plane_thread*)));
-    threads.back()->start();
-
-    threads.push_back(new Y_plane_thread<Word>(yitem,img, clamper, name));
-    connect(threads.back(), SIGNAL(finished(Volume_plane_thread*)), this, SLOT(addVP(Volume_plane_thread*)));
-    threads.back()->start();
-
-    threads.push_back(new Z_plane_thread<Word>(zitem,img, clamper, name));
-    connect(threads.back(), SIGNAL(finished(Volume_plane_thread*)), this, SLOT(addVP(Volume_plane_thread*)));
-    threads.back()->start();
-
-  }
-
-  template<typename T>
-  void switchReaderConverter(std::pair<T, T> minmax) { 
-    switchReaderConverter(minmax, typename boost::is_integral<T>::type()); 
-  }
-
-  template<typename T>
-  void switchReaderConverter(std::pair<T, T> minmax, boost::true_type) {
-    // IntConverter
-    IntConverter x = { minmax }; pxr_.setIC(x);
-  }
-
-  template<typename T>
-  void switchReaderConverter(std::pair<T, T> minmax, boost::false_type) {
-    // IntConverter
-    DoubleConverter x = { minmax }; pxr_.setFC(x);
-  }
-
-private Q_SLOTS:
-  //updates the msgBox
-  void update_msgBox()
-  {
-    static int nbPlanes =0;
-    nbPlanes ++;
-    msgBox.setText(QString("Planes created : %1/3").arg(nbPlanes));
-    if(nbPlanes == 3)
-    {
-      msgBox.hide();
-      nbPlanes = 0;
-    }
-  }
-  // Avoids the segfault after the deletion of an item
-  void erase_group()
-  {
-    Scene_segmented_image_item* img_itm = qobject_cast<Scene_segmented_image_item*>(sender());
-    if(img_itm)
-    {
-      if(group_map.contains(img_itm))
-      {
-        int iD = scene->item_id(group_map[img_itm].group);
-        CGAL::Three::Scene_group_item* group = qobject_cast<CGAL::Three::Scene_group_item*>(scene->item(iD));
-        if(!group)
-          return;
-        Q_FOREACH(CGAL::Three::Scene_item* child, group->getChildren())
-          scene->erase(scene->item_id(child));
-        scene->erase(iD);
-        group_map.remove(img_itm);
-      }
-    }
-    else
-    {
-      CGAL::Three::Scene_group_item* group_item = qobject_cast<CGAL::Three::Scene_group_item*>(sender());
-      if(group_item)
-      {
-
-        Q_FOREACH(CGAL::Three::Scene_item* key, group_map.keys())
-        {
-          if(group_map[key].group == group_item)
-          {
-            group_map[key].xitem = NULL;
-            group_map[key].yitem = NULL;
-            group_map[key].zitem = NULL;
-            group_map.remove(key);
-            break;
-          }
-        }
-      }
-    }
-  }
-  void connect_controls(int id)
-  {
-    CGAL::Three::Scene_item* sel_itm = scene->item(id);
-
-    if(!sel_itm || !group_map.contains(sel_itm)) //the planes are not yet created or the selected item is not a segmented_image
-    {
-      return;
-    }
-    controls c = group_map[sel_itm];
-    current_control = &group_map[sel_itm];
-    // x line
-    if(c.xitem != NULL)
-    {
-      Volume_plane_interface* x_plane = qobject_cast<Volume_plane_interface*>(c.xitem);
-      if(x_slider)
-        delete x_slider;
-      x_slider = new Plane_slider(x_plane->translationVector(), scene->item_id(x_plane), scene, x_plane->manipulatedFrame(),
-                                  Qt::Horizontal, x_control);
-      x_slider->setRange(0, (x_plane->cDim() - 1) * 100);
-      connect(x_slider, SIGNAL(realChange(int)), x_cubeLabel, SLOT(setNum(int)));
-      connect(x_plane, SIGNAL(manipulated(int)), x_cubeLabel, SLOT(setNum(int)));
-      connect(x_plane, SIGNAL(aboutToBeDestroyed()), this, SLOT(destroy_x_item()));
-      connect(x_slider, SIGNAL(realChange(int)), this, SLOT(set_value()));
-      x_slider->setValue(c.xvalue);
-
-      x_box->addWidget(x_slider);
-      x_box->addWidget(x_cubeLabel);
-    }
-    //y line
-    if(c.yitem != NULL)
-    {
-      Volume_plane_interface* y_plane = qobject_cast<Volume_plane_interface*>(c.yitem);
-      if(y_slider)
-        delete y_slider;
-      y_slider = new Plane_slider(y_plane->translationVector(), scene->item_id(y_plane), scene, y_plane->manipulatedFrame(),
-                                  Qt::Horizontal, z_control);
-      y_slider->setRange(0, (y_plane->cDim() - 1) * 100);
-      connect(y_slider, SIGNAL(realChange(int)), y_cubeLabel, SLOT(setNum(int)));
-      connect(y_plane, SIGNAL(manipulated(int)), y_cubeLabel, SLOT(setNum(int)));
-      connect(y_plane, SIGNAL(aboutToBeDestroyed()), this, SLOT(destroy_y_item()));
-      connect(y_slider, SIGNAL(realChange(int)), this, SLOT(set_value()));
-      y_slider->setValue(c.yvalue);
-      y_box->addWidget(y_slider);
-      y_box->addWidget(y_cubeLabel);
-    }
-    // z line
-    if(c.zitem != NULL)
-    {
-      Volume_plane_interface* z_plane = qobject_cast<Volume_plane_interface*>(c.zitem);
-      if(z_slider)
-        delete z_slider;
-      z_slider = new Plane_slider(z_plane->translationVector(), scene->item_id(z_plane), scene, z_plane->manipulatedFrame(),
-                                  Qt::Horizontal, z_control);
-      z_slider->setRange(0, (z_plane->cDim() - 1) * 100);
-      connect(z_slider, SIGNAL(realChange(int)), z_cubeLabel, SLOT(setNum(int)));
-      connect(z_plane, SIGNAL(manipulated(int)), z_cubeLabel, SLOT(setNum(int)));
-      connect(z_plane, SIGNAL(aboutToBeDestroyed()), this, SLOT(destroy_z_item()));
-      connect(z_slider, SIGNAL(realChange(int)), this, SLOT(set_value()));
-      z_slider->setValue(c.zvalue);
-      z_box->addWidget(z_slider);
-      z_box->addWidget(z_cubeLabel);
-    }
-  }
-//Keeps the position of the planes for the next time
-  void set_value()
-  {
-    current_control->xvalue = x_slider->value();
-    current_control->yvalue = y_slider->value();
-    current_control->zvalue = z_slider->value();
-  }
-
-  void destroy_x_item()
-  {
-    current_control->xitem = NULL;
-    x_slider->hide();
-  }
-  void destroy_y_item()
-  {
-    current_control->yitem = NULL;
-    y_slider->hide();
-  }
-  void destroy_z_item()
-  {
-    current_control->zitem = NULL;
-    z_slider->hide();
-  }
-
-};
-
-
-#include "Volume_planes_plugin.moc"
